+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support `OR` in queries
 - New API methods for Database:
   - async upsert(table, record, condition): Object
+- Add typing for Database and Query classes
 
 ## [1.1.3][] - 2021-05-25